+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - `feat`: wrap library for chain-style programming.
 - `feat`: ensure text environment shortcut to simplify testing.
+- `feat`: dart package test support both Dart and Flutter mode.
 
 ## 0.0.1